backend/data_parser.py
import os
import re
import pandas as pd
from docx import Document

DOCS_DIR = r"C:\Users\huiju919\Desktop\高血压糖尿病案例20260109\高血压糖尿病案例20260109"
OUTPUT_CSV = r"C:\Users\huiju919\Desktop\work_space\SmartDiag_Project\backend\cases.csv"

def extract_case_id(text):
    text = text.strip()
    if not text: return None
    if re.search(r'(?<![a-zA-Z])[A-Ea-e]\d+\s*(?:-|至|~)\s*[A-Ea-e]\d+', text): return None
    m = re.match(r'^([A-Ea-e])\s*(\d+)$', text, re.IGNORECASE)
    if m: return (m.group(1)+m.group(2)).upper()
    m = re.search(r'^[\[（\(]\s*([A-Ea-e]\d+)\s*[\]）\)]', text, re.IGNORECASE)
    if m: return m.group(1).upper()
    m = re.search(r'难度[：:]\s*([A-Ea-e]\d+)', text, re.IGNORECASE)
    if m: return m.group(1).upper()
    m = re.search(r'^([A-Ea-e]\d+)\s*案例', text, re.IGNORECASE)
    if m: return m.group(1).upper()
    m = re.search(r'案例.*?(?<![a-zA-Z0-9])([A-Ea-e]\d+)(?![a-zA-Z])', text, re.IGNORECASE)
    if m: return m.group(1).upper()
    return None

def parse_all_documents(docs_dir):
    all_cases = []

    for filename in os.listdir(docs_dir):
        if not filename.endswith(".docx") or filename.startswith("~$"): continue
        filepath = os.path.join(docs_dir, filename)
        print(f"\n📄 正在强行破译卷宗: {filename}")
        try:
            doc = Document(filepath)
        except Exception as e:
            continue

        current_case = None
        state = "idle"
        questions_lines = []
        answers_lines = []
        
        def save_current_case():
            if current_case and current_case["id"] != "UNKNOWN":
                current_case["questions"] = "\n".join(questions_lines).strip()
                current_case["standard_answers"] = "\n".join(answers_lines).strip()
                all_cases.append(current_case)

        for para in doc.paragraphs:
            text = para.text.strip()
            if not text: continue

            # 1. 抓取病人 ID (新建档案)
            case_id = extract_case_id(text)
            if case_id:
                save_current_case()
                level_char = case_id[0] 
                level_mapping = {'A': 1, 'B': 2, 'C': 3, 'D': 4, 'E': 5}
                current_case = {
                    "id": case_id,
                    "title": text[:30] + ("..." if len(text)>30 else ""),
                    "level": level_mapping.get(level_char, 1),
                    "initial_gold": 1000,
                    "kb_query": f"案例 {case_id}",
                    "bp": "未检查",
                    "blood_sugar": "未检查",
                    "temp": "未提及",
                    "auscultation": "未见异常",
                    "x_ray": "未见异常",
                    "questions": "",
                    "standard_answers": ""
                }
                state = "in_case"
                questions_lines = []
                answers_lines = []
                continue  

            if not current_case: continue

            # 2. 暴力抓取血压 (不管有没有mmHg，只要有类似 130/80 的格式就抓)
            if ("血压" in text or "BP" in text.upper()) and current_case["bp"] == "未检查":
                bp_match = re.search(r'(\d{2,3}\s*/\s*\d{2,3})', text)
                if bp_match: current_case["bp"] = bp_match.group(1) + " mmHg"

            # 3. 暴力抓取血糖 (不管有没有mmol/L，只要在血糖周围有小数就抓)
            if "糖" in text and current_case["blood_sugar"] == "未检查":
                sugar_match = re.search(r'(\d{1,2}\.\d{1,2})', text)
                if sugar_match: current_case["blood_sugar"] = sugar_match.group(1) + " mmol/L"

            # 4. 暴力追踪“问题”和“答案” (哪怕只写了“问题”两个字，或者用“五、”开头)
            if re.search(r'^(?:[四五六七八九十][、\.\s]*)?问题', text) or text == "问题" or "问题：" in text:
                state = "recording_questions"
                continue
            elif re.search(r'^(?:[四五六七八九十][、\.\s]*)?答案', text) or text == "答案" or "答案：" in text:
                state = "recording_answers"
                continue
            elif re.match(r'^十[一二]?[、\.\s]*', text) and state != "in_case":
                # 遇到大章节自动停止，防止串台
                state = "in_case"

            if state == "recording_questions":
                questions_lines.append(text)
            elif state == "recording_answers":
                answers_lines.append(text)

        save_current_case()
    return all_cases

def main():
    print(f"🚀 [自适应模式] 启动！全面清扫错漏数据！")
    all_extracted_cases = parse_all_documents(DOCS_DIR)
    
    if all_extracted_cases:
        df = pd.DataFrame(all_extracted_cases)
        os.makedirs(os.path.dirname(OUTPUT_CSV), exist_ok=True)
        df.to_csv(OUTPUT_CSV, index=False, encoding='utf-8-sig')
        print(f"\n🎉 完美收官！共计 201 个病例，漏网数据已全面修复！")
        print(f"📁 核心题库已覆盖: {OUTPUT_CSV}")

if __name__ == "__main__":
    main()