backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
httpx>=0.26.0
httpx-sse>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-docx>=1.1.0
pandas>=2.0.0
